langgraph/pyproject.toml
[project]
name = "aeris-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "Erik Peterson",email = "e@eriklp.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "langgraph (>=0.2.67,<0.3.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.60.1,<2.0.0)",
    "langchain (>=0.3.15,<0.4.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "newsapi-python (>=0.2.7,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.3"
types-requests = "^2.32.0.20241016"

